Change the structure of API views implementation
ECOM-2338
