+++ ecommerce/extensions/api/v2/views/__init__.py
@@ -1,0 +1,5 @@
+from rest_framework import viewsets, mixins
+
+
+class NonDestroyableModelViewSet(mixins.CreateModelMixin, mixins.UpdateModelMixin, viewsets.ReadOnlyModelViewSet):
+    pass
