+++ ecommerce/extensions/api/v2/views/baskets.py
@@ -1,41 +1,21 @@
-"""HTTP endpoints for interacting with Oscar."""
+"""HTTP endpoints for interacting with baskets."""
 import logging
 
-from django.conf import settings
-from django.contrib.auth import get_user_model
 from django.db import transaction
 from django.utils.decorators import method_decorator
-from oscar.core.loading import get_model, get_class
-from rest_framework import status, generics, viewsets, mixins
-from rest_framework.decorators import detail_route
-from rest_framework.exceptions import ParseError
-from rest_framework.permissions import IsAuthenticated, DjangoModelPermissions, IsAdminUser
+from rest_framework import status, generics
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
-from rest_framework_extensions.mixins import NestedViewSetMixin
-import waffle
 
-from ecommerce.core.constants import COURSE_ID_REGEX
-from ecommerce.courses.models import Course
 from ecommerce.extensions.analytics.utils import audit_log
-from ecommerce.extensions.api import data as data_api, exceptions as api_exceptions, serializers
+from ecommerce.extensions.api import data as data_api, exceptions as api_exceptions
 from ecommerce.extensions.api.constants import APIConstants as AC
-from ecommerce.extensions.api.exceptions import BadRequestException
-from ecommerce.extensions.api.permissions import CanActForUser
-from ecommerce.extensions.api.throttles import ServiceUserThrottle
 from ecommerce.extensions.checkout.mixins import EdxOrderPlacementMixin
-from ecommerce.extensions.partner.shortcuts import get_partner_for_site
 from ecommerce.extensions.payment import exceptions as payment_exceptions
-from ecommerce.extensions.payment.helpers import (get_processor_class, get_default_processor_class,
-                                                  get_processor_class_by_name)
-from ecommerce.extensions.refund.api import find_orders_associated_with_course, create_refunds
+from ecommerce.extensions.payment.helpers import (get_default_processor_class, get_processor_class_by_name)
 
 
 logger = logging.getLogger(__name__)
-
-Order = get_model('order', 'Order')
-Product = get_model('catalogue', 'Product')
-Refund = get_model('refund', 'Refund')
-User = get_user_model()
 
 
 class BasketCreateView(EdxOrderPlacementMixin, generics.CreateAPIView):
@@ -308,283 +288,3 @@ class BasketCreateView(EdxOrderPlacementMixin, generics.CreateAPIView):
             },
             status=status.HTTP_400_BAD_REQUEST
         )
-
-
-class OrderListView(generics.ListAPIView):
-    """Endpoint for listing orders.
-
-    Results are ordered with the newest order being the first in the list of results.
-    """
-    permission_classes = (IsAuthenticated,)
-    serializer_class = serializers.OrderSerializer
-
-    def get_queryset(self):
-        user = self.request.user
-        qs = user.orders
-
-        if user.is_superuser:
-            qs = Order.objects.all()
-
-        return qs.order_by('-date_placed')
-
-
-class OrderRetrieveView(generics.RetrieveAPIView):
-    """Allow the viewing of orders.
-
-    Given an order number, allow the viewing of the corresponding order. This endpoint will return a 404 response
-    status if no order is found. This endpoint will only return orders associated with the authenticated user.
-
-    Returns:
-        Order: The requested order.
-
-    Example:
-        >>> url = 'http://localhost:8002/api/v2/orders/100022'
-        >>> headers = {
-            'content-type': 'application/json',
-            'Authorization': 'JWT '  token
-        }
-        >>> response = requests.get(url, headers=headers)
-        >>> response.status_code
-        200
-        >>> response.content
-        '{
-            "currency": "USD",
-            "date_placed": "2015-02-27T18:42:34.017218Z",
-            "lines": [
-                {
-                    "description": "Seat in DemoX Course with Honor Certificate",
-                    "status": "Complete",
-                    "title": "Seat in DemoX Course with Honor Certificate",
-                    "unit_price_excl_tax": 0.0
-                }
-            ],
-            "number": "OSCR-100022",
-            "status": "Complete",
-            "total_excl_tax": 0.0
-        }'
-    """
-    permission_classes = (IsAuthenticated,)
-    serializer_class = serializers.OrderSerializer
-    lookup_field = AC.KEYS.ORDER_NUMBER
-
-    def get_queryset(self):
-        """
-        Returns a queryset consisting of only the authenticated user's orders.
-
-        This ensures we do not allow one user to view the data of another user.
-        """
-        return self.request.user.orders
-
-
-class OrderByBasketRetrieveView(OrderRetrieveView):
-    """Allow the viewing of Orders by Basket.
-
-    Works exactly the same as OrderRetrieveView, except that orders are looked
-    up via the id of the related basket.
-    """
-    lookup_field = 'basket_id'
-
-
-class OrderFulfillView(generics.UpdateAPIView):
-    permission_classes = (IsAuthenticated, DjangoModelPermissions,)
-    throttle_classes = (ServiceUserThrottle,)
-    lookup_field = 'number'
-    queryset = Order.objects.all()
-    serializer_class = serializers.OrderSerializer
-
-    def update(self, request, *args, **kwargs):
-        order = self.get_object()
-
-        if not order.is_fulfillable:
-            return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
-
-        logger.info('Attempting fulfillment of order [%s]...', order.number)
-        post_checkout = get_class('checkout.signals', 'post_checkout')
-        post_checkout.send(sender=post_checkout, order=order)
-
-        if order.is_fulfillable:
-            logger.warning('Fulfillment of order [%s] failed!', order.number)
-            return Response(status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-        serializer = self.get_serializer(order)
-        return Response(serializer.data)
-
-
-class PaymentProcessorListView(generics.ListAPIView):
-    """View that lists the available payment processors."""
-    pagination_class = None
-    permission_classes = (IsAuthenticated,)
-    serializer_class = serializers.PaymentProcessorSerializer
-
-    def get_queryset(self):
-        """Fetch the list of payment processor classes based on Django settings."""
-        return [get_processor_class(path) for path in settings.PAYMENT_PROCESSORS]
-
-
-class RefundCreateView(generics.CreateAPIView):
-    """
-    Creates refunds.
-
-    Given a username and course ID, this view finds and creates a refund for each order
-    matching the following criteria:
-
-        * Order was placed by the User linked to username.
-        * Order is in the COMPLETE state.
-        * Order has at least one line item associated with the course ID.
-
-    Note that only the line items associated with the course ID will be refunded.
-    Items associated with a different course ID, or not associated with any course ID, will NOT be refunded.
-
-    With the exception of superusers, users may only create refunds for themselves.
-    Attempts to create refunds for other users will fail with HTTP 403.
-
-    If refunds are created, a list of the refund IDs will be returned along with HTTP 201.
-    If no refunds are created, HTTP 200 will be returned.
-    """
-    permission_classes = (IsAuthenticated, CanActForUser)
-
-    def create(self, request, *args, **kwargs):
-        """ Creates refunds, if eligible orders exist. """
-        course_id = request.data.get('course_id')
-        username = request.data.get('username')
-
-        if not course_id:
-            raise BadRequestException('No course_id specified.')
-
-        # We should always have a username value as long as CanActForUser is in place.
-        if not username:  # pragma: no cover
-            raise BadRequestException('No username specified.')
-
-        try:
-            user = User.objects.get(username=username)
-        except User.DoesNotExist:
-            raise BadRequestException('User "{}" does not exist.'.format(username))
-
-        refunds = []
-
-        # We can only create refunds if the user has orders.
-        if user.orders.exists():
-            orders = find_orders_associated_with_course(user, course_id)
-            refunds = create_refunds(orders, course_id)
-
-        # Return HTTP 201 if we created refunds.
-        if refunds:
-            refund_ids = [refund.id for refund in refunds]
-            return Response(refund_ids, status=status.HTTP_201_CREATED)
-
-        # Return HTTP 200 if we did NOT create refunds.
-        return Response([], status=status.HTTP_200_OK)
-
-
-class RefundProcessView(generics.UpdateAPIView):
-    """
-    Process--approve or deny--refunds.
-
-    This view can be used to approve, or deny, a Refund. Under normal conditions, the view returns HTTP status 200
-    and a serialized Refund. In the event of an error, the view will still return a serialized Refund (to reflect any
-    changed statuses); however, HTTP status will be 500.
-
-    Only staff users are permitted to use this view.
-    """
-    permission_classes = (IsAuthenticated, IsAdminUser,)
-    queryset = Refund.objects.all()
-    serializer_class = serializers.RefundSerializer
-
-    def update(self, request, *args, **kwargs):
-        APPROVE = 'approve'
-        DENY = 'deny'
-
-        action = request.data.get('action', '').lower()
-
-        if action not in (APPROVE, DENY):
-            raise ParseError('The action [{}] is not valid.'.format(action))
-
-        refund = self.get_object()
-        result = False
-
-        if action == APPROVE:
-            result = refund.approve()
-        elif action == DENY:
-            result = refund.deny()
-
-        http_status = status.HTTP_200_OK if result else status.HTTP_500_INTERNAL_SERVER_ERROR
-        serializer = self.get_serializer(refund)
-        return Response(serializer.data, status=http_status)
-
-
-class NonDestroyableModelViewSet(mixins.CreateModelMixin, mixins.UpdateModelMixin, viewsets.ReadOnlyModelViewSet):
-    pass
-
-
-class CourseViewSet(NonDestroyableModelViewSet):
-    lookup_value_regex = COURSE_ID_REGEX
-    queryset = Course.objects.all()
-    serializer_class = serializers.CourseSerializer
-    permission_classes = (IsAuthenticated, IsAdminUser,)
-
-    def get_serializer_context(self):
-        context = super(CourseViewSet, self).get_serializer_context()
-        context['include_products'] = bool(self.request.GET.get('include_products', False))
-        return context
-
-    @detail_route(methods=['post'])
-    def publish(self, request, pk=None):  # pylint: disable=unused-argument
-        """ Publish the course to LMS. """
-        course = self.get_object()
-        published = False
-        msg = 'Course [{course_id}] was not published to LMS ' \
-              'because the switch [publish_course_modes_to_lms] is disabled.'
-
-        if waffle.switch_is_active('publish_course_modes_to_lms'):
-            access_token = getattr(request.user, 'access_token', None)
-            published = course.publish_to_lms(access_token=access_token)
-            if published:
-                msg = 'Course [{course_id}] was successfully published to LMS.'
-            else:
-                msg = 'An error occurred while publishing [{course_id}] to LMS.'
-
-        return Response({'status': msg.format(course_id=course.id)},
-                        status=status.HTTP_200_OK if published else status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-
-class ProductViewSet(NestedViewSetMixin, NonDestroyableModelViewSet):
-    queryset = Product.objects.all()
-    serializer_class = serializers.ProductSerializer
-    permission_classes = (IsAuthenticated, IsAdminUser,)
-
-
-class AtomicPublicationView(generics.CreateAPIView, generics.UpdateAPIView):
-    """Attempt to save and publish a Course and associated products.
-
-    If either fails, the entire operation is rolled back. This keeps Otto and the LMS in sync.
-    """
-    permission_classes = (IsAuthenticated, IsAdminUser,)
-    serializer_class = serializers.AtomicPublicationSerializer
-
-    def get_serializer_context(self):
-        context = super(AtomicPublicationView, self).get_serializer_context()
-        context['access_token'] = self.request.user.access_token
-        context['partner'] = get_partner_for_site(self.request)
-        return context
-
-    def post(self, request, *args, **kwargs):
-        return self._save_and_publish(request.data)
-
-    def put(self, request, *args, **kwargs):
-        return self._save_and_publish(request.data, course_id=kwargs['course_id'])
-
-    def _save_and_publish(self, data, course_id=None):
-        """Create or update a Course and associated products, then publish the result."""
-        if course_id is not None:
-            data['id'] = course_id
-
-        serializer = self.get_serializer(data=data)
-        is_valid = serializer.is_valid(raise_exception=True)
-        if is_valid:
-            created, failure, message = serializer.save()
-            if failure:
-                return Response({'error': message}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-            else:
-                content = serializer.data
-                content['message'] = message if message else None
-                return Response(content, status=status.HTTP_201_CREATED if created else status.HTTP_200_OK)
